move file row into separate folder
